Temperature

+++ DeepSeaHunter/DeepSeaHunter/Assets/Scripts/General/ITangle.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface ITangle
 {
-    void stateTangled(int tangleMod);
-    void stateUntangled(int tangleMod);
+    void stateTangled(float tangleMod);
+    void stateUntangled(float tangleMod);
 }
